+++ value/dict/combined.go
@@ -10,7 +10,7 @@ import (
 )
 
 func Addr(opts ...DictOpt[netip.Addr]) value.EmptyConstructor {
-	return New(contained.Addr(), opts...)
+	return New(contained.NetIPAddr(), opts...)
 }
 
 func AddrPort(opts ...DictOpt[netip.AddrPort]) value.EmptyConstructor {